Revert "Create hooks for all logic in pages/locale/index.tsx"
